Model panels as a real Room entity with add/rename/delete/reorder
Panels previously had no stable identity — just a CSV name list in
DataStore keyed by list position, with per-panel HAL actions encoded
the same fragile way. Deleting a panel mid-list silently reassigned
favorites/HAL config to the wrong panel and never cleaned up orphaned
rows. Panels are now a Room entity with FK cascade delete on
favorite_apps, giving real add/rename/delete/reorder UI (EditPanelsDialog)
and safe cleanup when a panel is removed. Magic-button config is now a
single global setting shared across panels instead of per-panel.

Also fixes a couple of build warnings/errors surfaced along the way:
a wrong awaitFirstDown import, a FlowPreview API usage without opt-in,
a deprecated LocalLifecycleOwner import, and an AGP/compileSdk mismatch
warning.

+++ app/src/main/java/com/ykatchou/ylauncher/data/db/FavoriteDao.kt
@@ -32,7 +32,7 @@ interface FavoriteDao {
     suspend fun deleteAll()
 
     @Query("DELETE FROM favorite_apps WHERE panelId = :panelId")
-    suspend fun deleteByPanel(panelId: Int)
+    suspend fun deleteByPanel(panelId: Long)
 
     @Query("SELECT COUNT(*) FROM favorite_apps")
     suspend fun count(): Int
